--- ORIG-fn-prevalence-predictor/function/preprocess_params.py
@@ -1,5 +1,0 @@
-from preprocess_helpers import write_temp_from_url_or_base64, required_exists
-
-
-def preprocess(params: dict):
-    type(params['request_parameters']['threshold']) in [float, int]
