+++ fn-prevalence-predictor/preprocess_params.py
@@ -1,7 +1,5 @@
 from preprocess_helpers import write_temp_from_url_or_base64, required_exists
 
-# Receives the params dict
-# Throw an error any time a test for a param fails.
-# Otherwise, return nothing
+
 def preprocess(params: dict):
-    raise ValueError('Need to implement `preprocess` steps in function/preprocess_params.py')
+    type(params['request_parameters']['threshold']) in [float, int]
