+++ fn-prevalence-predictor/requirements.txt
@@ -1,0 +1,8 @@
+git+git://github.com/disarm-platform/disarm-gears.git@master
+numpy
+matplotlib
+pandas
+sklearn
+pygam
+descartes
+requests
